Fix dependency on googlechart for graphviz
The initial commit to allow graphviz images in documentation used
a sphinx extension relying on googlechart (which just sends the .dot files
to a google API and let's them render it). It turns out google's
implementation of the .dot renderer is quite out of date, and there's a
better module we can include which doesn't rely on this third party service
anyway.

Change-Id: I2bd10dc68d1d108af6bbf7f0e3871642f294eb8c