Octavia v0.5 component design
This design represents the overall layout of components that will
make up the first major milestone release of Octavia (v0.5). The
intent here is to have a scalable service delivery layer, to get
communication with external components right, and not yet worry
about making the command and control layer scalable.

Change-Id: I09ab7185683b66ec2567119616bc54cf9d8cc000