Added vrrp_ip and ha_ip to amphora models
vrrp_ip will typically be the ip through which the amphora communicates with
its vrrp peer.

ha_ip will typically be the ip that is shared between the amphorae.

Since it is possible each amphora may have a different ha_ip and vrrp_ip,
it makes sense to add them to the model/table.  Adding them now because the
network driver will be assigning the values.

Changed the network driver spec and base class to mention the plug_vip method
should return a list of amphoras.

Change-Id: I04a97caf00bc6fa25f94e6470d3ed7da48880ae6