Fix ups so pbr and setup works
Moved octavia/bin directory to octavia/octavia/cmd and renamed octavia-api to
api.py.  This follows the new strucutre most openstack projects are doing.
Also implemented octavia-api being built when doing setup install.

Had to remove the octavia/tests/specs symlink because pbr did not like it and
would not allow anything to be built. Also found out the test_titles test was
not being run on the specs, so that has been added as a separate tox
environment.

Change-Id: I92f2bc6eaa346983f7a2479dc9fa43d2040f2cd5