Added update_vip method to network driver
There needed to be a method to update the security group rules whenever
a listener is added or removed.  The update_vip method will not update those
rules based on what listener's are present.

Also changed the allocate_vip method to take in a load_balancer instead of
port_id, network_id, and/or ip_address.   The reason for this is some driver
implementations may just want the vip to be the IP directly on the amphora.
The previous signature did not allow this.

Closes-Bug: #1453609
Closes-Bug: #1453610

Change-Id: Ie5765c231c6f6ba45042db9b111e6814cf50c465