Adding methods to network driver to get more info
The network driver needs the ability to retrieve specific information about
subnets and ports and pass that information to other components.  Having info
about the subnet gives us cidr and gateway data that is valuable when setting
up the amphora's routes for the separation of tenant data and management data.
Because of this need, methods to retrieve network, subnet, and port details
has been added.

This also adds a another data model AmphoraNetworkConfig that contains subnet
and port information for the vip, vrrp, and ha networks on an amphora.  A
network task has been added that builds this data structure by calling the new
methods added to the network driver.

Change-Id: Ica912337c7d52659a9733096878664f734f27b00