Decouple amp driver from network using new models
The amphora driver's post_vip_plug method required network specific information
about the amphora to complete implementations.  It was not accepting anything
other than a loadbalancer object.  Now it takes a dictionary of
AmphoraNetworkConfig objects that is keyed of amphora id.  Each instance
contains network specific information required to set up the correct routing
on the amphora.

This patch sets up the routing correctly to solve the vip blackholing issue but
only for the ssh_driver.  The argument has only been added to the post_vip_plug
method of the rest_api_driver but will need to be updated to handle this new
information and to also fix the vip blackholing problems.

Change-Id: I17ce89b6c050a2a36e0a802920e2dedb063f615b
Closes-Bug: #1453951