Fix member creation when retrieving network
When a member is created there must be some calculation to determine what
networks the amphorae need to be plugged into.  This calculation needs to
retrieve network information.  How it retrieves information was based on an
old version of the network driver and was missed during the network driver
decoupling refactor.  The calculate delta method now calls the get_subnet
method of the member's subnet_id attribute.

Change-Id: I65da3ae5f02b991fb6aad5a267dc02b0572b35fd