Exclude virtual interfaces for collecting stats
The /details api provides details about the amphora VM.
It reads the interface's statistics file
/sys/class/net/{interface}/statistics/{type}_bytes
to collect all the stats. While doing so it tries to read the
stat files of the virtual interfaces like eth0:1 also,
which are not available and throws an exception. This
fix excludes virtual interfaces while collecting the
stats by checking whether the interface name has : in it

Closes-Bug:1466688
Change-Id: Ie2786f18b96881ba600a338bcad916677709a5b3