Check for port ownership on vip deallocation
The code was checking the tenant to determine if octavia had created the
vip port or not. If it did create the vip port then it would delete it, if not
it would not delete the port. Since neutron-lbaas merged a change that set
the device_owner appropriately on the port, we can now check device_owner for
ownership. This also adds octavia as the device_owner of the vip port if
it does create it and sets the device_id to the load balancer's id.

Change-Id: I12adf2824c77f633be7499f120e6f5b3a9c326ef