Do not remove egress sec group rules on plug vip
When a vip is plugged or a listener is added, all security group rules that
did not have the current configuration of ports were deleted.  This included
the egress rules that by default allow all protocols and ports.  Now, egress
rules are ignored and maintained so connection to amphorae can be preserved.

Change-Id: I9ff2772f9c3e12a5099b7b57cd64f74da5aa35ae
Closes-Bug: #1488281