Rest API plug network interface handling
On a plug network the rest api code was trying to bring up an interface
appended by :0 which is only meant for multiple ips on the same link.  When
post network plug is called on the amphora driver, this is because a member
got added that is on a different subnet.

This change also checks to see if there are any network changes before calling
post_plug_network of the amphora driver.  It didn't make sense to call this
method if there weren't any networks plugged.

Change-Id: I76548a8c82a7a495f909b5b8ac3932a661731293