Fix issues uncovered by neutron lbaas tempest tests
Updating a listener caused issues because SNI was not being appropriately
handles on a listener update.  This has been "fixed" but another patch will be
needed to make sure that every listener update does not clear out the list. I
think it is best to leave that for another review though.

This also fixes the bug where security groups were not getting cleaned up when
a load balancer is deleted.  Since neutron does not synchronously remove a
security group from a port, a retry mechanism had to be used and thus more
config options to tune the interval and max attempts.

Change-Id: I0434b8ced144ab08413b91569bd008295ef1784e
Closes-Bug: #1464953