Fix a typo in the .coveragerc file
Coverage now aborts the run when the configuration file has unknown
settings.  This patch fixes a typo in the .coveragerc file that
would cause coverage jobs to fail.

A global requirements change will also be necessary to exclude
version 4.0 of coverage as it causes tests to fail that
do not fail outside of coverage or with earlier versions.
Tracked here: https://review.openstack.org/#/c/228581/

Global requirements changes come in as separate patches, so
I am not creating a dependency. 

Change-Id: Ifbbf2096303770e393e666e3562c83d863e82f15