REST Amp Agent: Compare lower case MAC Addresses
When the post_plug_vip method is called on the rest amphora driver,
the agent on the amphora running the rest server attempts to find the
interface by the MAC address passed in through the API. On some OSes
the interfaces returned by the OS has the mac address in all lower case,
whereas the MAC address passed in through the API is in upper case.
Just lower() both and then compare.

Closes-Bug: #1507881
Change-Id: I4ddfd25911d31285513009432f80da080196e8b2