Fix a problem building spare amphora
When amphora a being built via the housekeeping manager spares
pool management the amphora records in the database are not
being properly updated with the nova details. This results in
the amphora not having a lb_network_ip in the database.
This patch makes sure the amphora database record is updated.

Change-Id: I4a364b6fc631d19f4a93f9fe7c5348a2c9617790
Closes-Bug: #1515009