Fix api_server agent test that broke due to a dep change
The test was slightly incorrect to begin with, but the issue didn't
become apparent until the oslo.log update from 1.13.0 to 2.0.0 for
whatever reason. open() raises IOError() not Exception() so it was not
being handled properly.

Also taking this opportunity to standardize the use of builtins to
match our other tests.

Change-Id: Idab637b09e8a437333f810f82f45b3e9c027dfdd