Expose project_id in octavia api
The project_id (formerly tenant_id) was not actually allowed as an
attribute of any entity on create nor would it be returned as details
of an entity.  project_id will be used in the future and is just
overall good to have.

This also accepts request headers to set the project_id if they are passed in
and no tenant_id is provided in the body.

Change-Id: I6408b7863409870b502de03af35692b065b2c6dc