Setup health manager networking for devstack
When running the health manager on single host devstack machine, the
hosts IP cannot simply be used as if the health manager is running
on a host by itself.  The solution here is to create a neutron port
on the lb-mgmt-net which sets up necessary ovs ports and also
reserves an IP.  Then the script creates an ovs port for and sets
and interface with the information the neutron port create allocated.

There are some flaws with this but its a starting point and can and
probably will be improved on in the future.  However it works for
now.

Change-Id: Ic3d3d1d63a5cc352c5fc00dea58bb16915754a7c
Closes-Bug: #1490033