Make Barbican the default certificate manager
Local certificate manager is broken and should probably be removed,
since the effort was never taken to make it functional (adding
additional fields for taking raw certificate data in the API).
Barbican should be the default as it is the functional option.

Also taking this opportunity to clean up the config and local.conf
defaults to match reality.

Change-Id: Icf1e89838b761e7d2926c8eb7532404b4c2f19dd