Case-sensitive string checking for Enum fields
This patch introduces stricter checking on fields that are in
lookup tables, but are effectively Enum fields. Because we do a
lot of case-sensitive string comparisons against these fields in
various places throughout the code tree, but because SQL does not
to case-sensitive comparisons by default, it's important that we
are case-sensitive in the API for these fields.

Change-Id: Ib23796754ca54f69bb60c79c75804a5e7ee26e29
Closes-Bug:1534217