[stable/liberty] Checkout stable/liberty version for amphora image
Currently when an amphora image is being built by diskimage-builder
it always checks out master branch.

Master version of amphora-agent has API that is incompatible
with stable/liberty version of client (haproxy rest_api_driver),
upload_haproxy_config method signature is different, for example.

Due to that (and the fact that API version wasn't bumped after changing
the API) attempts to upload haproxy config on amphora result in
400 "Unknown action: haproxy" as the request gets routed to
start_stop_listener instead of upload_haproxy_config. As a result
it is impossible even to create listeners with stable/liberty
installations.

The easiest way to fix this is to force diskimage-builder to fetch
the stable/liberty branch.

Change-Id: Ie688a45af14fd15a8304fd42a12a85dd05c43892
Closes-Bug: #1535341