Let drivers decide to use region or not
Passing region, or any other service lookup, authentication, authorization to
the drivers from the controller worker doesn't make much sense.  The driver
implementation should just pull it from the config if the driver needs it.
Otherwise it doesn't matter.  The controller technically should know nothing
about how the driver communicates out, or any setup information like that.
It should be up to the driver to decide that.

Change-Id: Ia6ffcbbe85b97b1ec278ed1ae2d48ca7cedf2834
Closes-Bug: 1539433