Add L7 api - policies
This commit adds the API necessary for manipulating L7 policies
in Octavia as well as unit tests of the API. It is one in a chain
of commits designed to keep the size of each individual commit
manageable / reviewable. API for L7 Rules, jinja template, and
documentation updates will come in later commits.

Partially-Implements: blueprint lbaas-l7-rules
Partially-Implements: blueprint layer-7-switching
Change-Id: Ib8b18beac83ca46d0fae30a76b72ef0e729696b7