Fix / and /v1 API uris
An update in the past added @staticmethod decorators to everything that didn't
use self in the function body.  Thise broke the way pecan calls the API
methods because pecan calls each method independent of the instance.  Pecan
instead passes the instance as the first argument.  With staticmethod not
having this first argument, it broke.

Change-Id: I19e30ecae3c1d49e569950b2675a73b2fbc1f942