Add L7 api - rules
This commit adds the API necessary for manipulating L7 rules
in Octavia as well as unit tests of the API. It is one in a chain
of commits designed to keep the size of each individual commit
manageable / reviewable. Jinja template and documentation updates
will come in later commits.

Partially-Implements: blueprint lbaas-l7-rules
Partially-Implements: blueprint layer-7-switching
Change-Id: Iabc3ac44bbb4c3230eb02f978300caabc0c25cc5