Update data model conversion to be more complete
Prior to this patch, the conversion from database model to
data model tried various recursive strategies based on object
class tracking, aimed at avoiding too many recursive back-references
to other models in the object tree in order to avoid an infinite
recursion. While this worked reasonably well, it resulted in
data model trees that would become increasingly inaccurate if
too many links were traversed away from the original object from
which the tree was generated (ie. the defacto root of the tree
that had been generated).

This problem became worse when the tree became a graph with the
introduction of the shared pools code, and the work-around for
the graph inaccuracy problem greatly increased the inefficiency
of generating the graph. (It's an O(n^2) or worse algorithm.)

This patch introduces a new strategy for generating the data model
graph based on tracking unique ID of the objects in the graph, and
making all back-references complete. By generating the graph this
way, not only do we have a complete and accurate data model graph
which should remain accurate no matter how many object links are
traversed (ie. this is a true graph and not an approximation of
a graph using a tree), but we have increased the efficiency of the
graph generation algorithm to O(n).

Partial-Bug: 1544851
Change-Id: I68c2ff0e293e3a1ea4e7b0fd38cf6d0d5cd4ec12