Add listener stats API
Octavia has been gathering listener statistics for a long time,
but so far has never exposed this via the API. Further, our
documentation has shown that we should have had an API interface to
the listener stats this whole time. This commit adds this interface,
as described in the API documentation.

Change-Id: I4848bb31feb03ab0599a26f600951518f2c41254
Closes-Bug: 1543356