Fix deleted amphora expiration check
During a normal failover, the housekeeping manager will attempt
to clean up database records having to do with the dead amphora based
on the last health check of said amphora. However, at least some of
the time, the amphora health manager would have already cleaned up
the health record, meaning that the housekeeping manager gets a
None response when it attempts to check the deleted amphora health.
This could lead to a failure in the dead amp cleanup process, possibly
gumming up failover or other tasks along the way.

This commit fixes this check such that if no health record exists
for the deleted amphora, it is assumed the amphora is expired and
all record of it can be deleted from the database.

Change-Id: Icded7127d7817d8c71a5fcb90c95775a2bf5f58c
Closes-Bug: 1544427