Fix amphora failover after health monitor start
The health monitor consists of two daemons: One receives heartbeats
from amphorae and updates the database, the other scans the database
for amphorae that haven't checked in in a while. If the health
monitor is down for a while, all the amphorae will seem stale in
the database once it's started again.

This patch changes the behavior of the health_check such that amphorae
have time to check in after the health monitor start before the
health monitor daemon decides they need to be failed over.

Change-Id: I0599d214c645547b3248f73f6bb3380653de0585
Closes-Bug: 1532623