Trivial: cleanup unused conf and log variables
Basic clean up work.
In some files, oslo_config.cfg and oslo_log.log  has been imported
but not used. So remove them.

I noticed that there are a lot of unused conf variables in nova and
there is a patch to clean them up:
https://review.openstack.org/#/c/280068/
So I wondered if this was a thing in octavia as well.
Turns out it is. :- )

If have to say why not keeping them, I think it should be making
code clean and avoiding redundant compilation.

Change-Id: I59b3dcef9143db2dbaae0c9c51f4e098ddcc16e3