Fix LB/Listener status updates for HM/Member
Recent scenario tests are showing create operations on
pool-subordinate objects (like Member and Health Monitor) appear
to be returning too quickly with statuses that indictate the
operation is done. This has lead to intermittent problems of tests
failing because the load balancer was not in an expected state.

Our hypothesis is that the aforementioned operations are not
setting the load balancer status to PENDING_UPDATE as they should.
This patch fixes this problem, ensuring that PENDING_UPDATE status
is set on the load balancer and listener status correctly prior to
the execution of the creation taskflow, as well as whenever pool or
pool-subordinate objects are updated and deleted. This should resolve
the related intermittent scenario test failures.

In addition to the above, I have re-organized some of the API methods
to remove some code duplication between various API controller, and
re-organized several of the API functional tests for health monitors
and members to catch the above problem.

Change-Id: I5e3a1eabb76af3b913e3edbe95a3b14c3c40244b
Closes-Bug: 1548588