Fix default_pool_id type in l7policy API
The L7Policy API type checker presently lists the redirect_pool_id
parameter as StringType(). While this is not incorrect, it is more
correct to call it a UuitType().

In addition to the above change, I removed unnecessary string length
checking on string response parameters.

No tests needed updating, as we already have good test coverage of the
validations we do on the redirect_pool_id in the L7Policy API.

Change-Id: I2bf70b51c6bc0949165f8af31c7d47c3407df08b
Closes-Bug: 1549013