Fix admin-state-up
While updating the admin-state-up of LB, Listener, Pool,
or member the update fails due to the wrong order of the
positional parameters.

This leads to SQL syntax error and eventually
Loadbalancer gets stuck in PENDING_UPDATE provisioning
status.

The haproxy jinja was not taking into account if objects
were enabled or not.

This patch fixes the above issue.

This patch also adds the OpenFixture class from neutron
to enable safer mocks of os.open.  This was needed for the
test of this code.

Co-Authored-By: Michael Johnson <johnsomor@gmail.com>
Closes-Bug: #1547123
Closes-Bug: #1549036
Closes-Bug: #1549186
Closes-Bug: #1552098
Change-Id: Id5733cc8e26ec7092b821c2c18c41696d6789265