Removes port_detach code that was mistakenly added
A previous patch [1] mistakenly added a port detach call to the failover
flow.  It didn't cause failures due to a type-o in the arguments to the
call, so it would passively fail with a log entry.  However, if someone
fixed this it would cause failover to fail due to the port getting
deleted during the detach.
This patch removes that code.

[1] https://review.openstack.org/#/c/295475/

Change-Id: I1eb1ad551636d5cd9253690a59e0e88ca1d238b5