Allow deployment with pre-generated ssh keys and certificates
This patch is part of a series that will allow the deployment
of a simple multinode octavia devstack, where the certs and
ssh keys are uniform over all the controllers.

Orchestrating the deployment of the same ssh keys or certificates
becomes difficult without making use of tools like ansible otherwise.

The provided certificate is valid for 10 years.

Change-Id: I45e8e54d58a725281636a61fdb769265de50d9ba