Fixed unit test for amphorae backend utils
In the function test_get_network_interface_file:
first we check path returned when 'agent_server_network_dir' is defined.

In the tested function the code of taking 'agent_server_network_dir'
is unreachable if there is a non-empty sting in 'agent_server_network_file'.
Hence, to test this case we should first set 'agent_server_network_file'
to None.

Change-Id: Ifb5951111585ee871e1956575d5434e09da8f8cf
Closes-Bug: 1585919