Fix alembic migration on MySQL 5.7
Presently, attempting to run alembic migrations for Octavia on a host
using MySQL 5.7 as the database back-end fails with an error about a
nullable field in a primary key. This commit fixes the offending
migration and introduces a second, similar migration so that any
existing installation using MySQL 5.6 or older also have this field set
to be non-nullable.

Change-Id: Iae3d4f627c1124b1d913aab283e6e19d9c0f0dbb
Closes-Bug: #1590145