Decrease default resource expiry age in DevStack setup
DevStack is used mostly for development purposes that
often involve frequent creation and deletion of resources.
That is why having a lot of stale resources like amphorae and
load balancers is quite confusing.

Keeping them around for a week makes sense in production but in
DevStack deployments it is reasonable to decrease expiry age
to 1 hour by default.

Change-Id: I1b78963256467175537caffd46a863da9f5fb018