Cache neutron extension lookup state
I don't believe that neutron extensions are loaded/unloaded on the fly,
so there should be no reason to check them every time (this adds
additional API call overhead). It is trivial to cache them instead.
If people think this is a bad idea for some reason I have not
considered, feel free to -2 (maybe my assumption is incorrect).

Change-Id: Ifa08bea3187eb1f603a28c7faf33dfbaff6f6afe